+++ src/lib.rs
@@ -1,2 +1,22 @@
-/// データから指定したパターンに合うデータを取り出す
+//! # MyParserProject
+//!  
+//! parserという題材のもとやってる学習目的のrepo  
+//! 将来的には HTTP Parser や 独自プロトコルの作成 、画像ファイルパーサー や コマンドライン、言語の作成を目指したい。  
+//!
+//! ## My Rust Base Repositoly
+//! rustで開発する上で、大体必要そうな装備を備えたリポジトリ
+//!
+//! ### actions
+//! CI/CD促進のために設置。　テストやドキュメント作成を自動化
+//!
+//! ### tests
+//! テスト開発も行なっていきたい
+//!
+//! ### errorの実装
+//! 独自でエラーを実装してる
+//!
+//! ### ブランチモデルを弱い採用
+//! 一般的なブランチモデルを参考に、ブランチ管理を行う。
+//! `main` `dev` 等
+
 pub mod parser;
