+++ src/parser.rs
@@ -1,23 +1,26 @@
-/// 単体でパーサーとして動作する、もっとも基本的なパーサー群。  
-///
-/// baseパーサーは、入力された文字列を左側からパースします。
-/// ## 成功時
-/// パースに成功したら、適切な型に変換されて戻されます。
-/// ## 失敗時
-/// パースに失敗したら、[crate::parser::Error] が返されます。
-/// ## 共通
-/// パースされなかった文字列は、パースされた文字列を除いた残りの文字列として返されます。
-///
-pub mod base;
-
-/// 複数のパーサーを組み合わせて動作するパーサー
+//! 配列・文字列から指定した条件で解析を行う機能を提供する
+//!
+//! パーサー（構文解析器）とは、文字列やバイト列のスライスから、指定した条件で解析を行うためのアルゴリズムのことです。
+//! このモジュールでは、トップダウン構文解析として、配列の先頭から解析を行うパーサーを提供します。
+//!
+//! # パーサーの型について
+//! 基本的なパーサーの構造は、 [Parser] トレイトに従います。  
+//! [combinator]モジュールで、パーサーを組み合わせて動作するパーサーを作成できます。  
 pub mod combinator;
+pub mod str_parser;
 
 // エラー
 mod error;
 pub use error::*;
 
-/// パーサー関数をトレイトオブジェクト化
+/// パーサー関数をトレイトオブジェクト化します。
+///
+/// 引数に指定されたスライスの先頭から、指定された条件で解析をし、一致した場合は結果を返却します。
+///
+/// # 戻り値について
+/// `(I, R)`のタプル形式です。
+/// - I: パース後に残った部分が入ります
+/// - R: パースした結果が入ります。結果の型式については、パーサーによって異なります。
 pub trait Parser<I, R>: Fn(I) -> (I, R) + Clone {}
 // Fn(I) -> (I, Result<T,E>) を Parser<I, T, E> として使えるようにする
 impl<I, R, F> Parser<I, R> for F where F: Fn(I) -> (I, R) + Clone {}
