--- src/parser/base/mod.rs
@@ -1,16 +1,0 @@
-use super::*;
-
-mod num;
-pub use num::*;
-
-mod char;
-pub use char::*;
-
-mod string;
-pub use string::*;
-
-mod trimer;
-pub use trimer::*;
-
-mod none;
-pub use none::*;
