+++ src/parser/combinator/map.rs
@@ -26,7 +26,7 @@ mod tests {
     fn map_2x_test() {
         // numパーサーで数値を取得し、2倍した結果を取得するテスト
         let base = "10x";
-        let parser = base::num.map(|x| x.map(|i| i * 2));
+        let parser = str_parser::num.map(|x| x.map(|i| i * 2));
         let (i, r1) = parser(base);
         // パース結果
         assert_eq!(r1, Ok(4));
@@ -38,7 +38,7 @@ mod tests {
     fn map_error_change_test() {
         // numパーサーのエラーを0に変換する
         let base = "x123";
-        let parser = base::num.map(|x| x.unwrap_or(0));
+        let parser = str_parser::num.map(|x| x.unwrap_or(0));
         let (i, r1) = parser(base);
         // パース結果
         assert_eq!(r1, 0);
