+++ src/parser/combinator/mod.rs
@@ -1,3 +1,4 @@
+//! 複数のパーサーを組み合わせて動作するパーサー
 use super::*;
 
 mod concat;
