+++ src/parser/error/errorkind.rs
@@ -1,8 +1,14 @@
-/// [super::Error] で使用する、ジャンル分けされたエラー原因を格納するenum
+/// エラーの種類を表すenumです。
+///
+/// - [Error::kind][super::Error::kind] で使用されます。
+/// - 各項目の詳しい解説は、エラーを返す関数の項目を参照してください。
 #[derive(PartialEq, Debug)]
+// 各Kindの解説は、実際にエラーを返す関数の項で説明
 pub enum ErrorKind {
-    /// パースできない時
+    /// [num][crate::parser::str_parser::num] パーサーのエラー
     ParseNumError(std::num::ParseIntError),
+    /// [char][crate::parser::str_parser::char] パーサーのエラー
     ParseCharError,
+    /// [string][crate::parser::str_parser::string] パーサーのエラー
     ParseStringError,
 }
