+++ src/parser/str_parser/string.rs
@@ -2,19 +2,25 @@ use super::*;
 
 /// 文字列でパースする
 ///
+/// 引数に入れた文字列を条件に解析する [super::Parser] を**作成**します。
+///
+/// # パーサーの仕様について
+/// - 先頭から解析を行います。
+/// ## 成功時
+/// - 条件の文字列が解析できたら、その文字列が結果として返されます。
+/// ## エラー時
+/// - [エラー][super::error::Error]が返却されます。
+///   - [kind][super::error::Error::kind]が [ParseStringError][super::ErrorKind::ParseStringError] になります。
+///
 /// # Example
 /// ```rust
-/// use crate::parser::base::string;
+/// use crate::parser::str_parser::string;
 ///
 /// let input = "hello_world";
 /// let (rest, result) = string("hello".to_string())(input);
 /// assert_eq!(result, Ok("hello"));
 /// assert_eq!(rest, "_world");
 /// ```
-///
-/// # Error
-/// kindが [crate::parser::ErrorKind::ParseStringError] になります。
-///
 pub fn string<'a>(s: String) -> impl Parser<&'a str, Result<String, Error>> {
     move |i: &'a str| todo!()
 }
